pyproject.toml
[project]
name = "prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.1.0",
    "chromadb>=1.4.0",
    "docx2txt>=0.9",
    "ollama>=0.1.0",
    "langchain>=1.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.12.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
